+++ test/Scripts.Integration.Test/integration-test.ps1
@@ -87,7 +87,7 @@ If ($Rebuild -or -not(Test-Path -Path $(GetNewProjectBuildPath))) {
     If ("iOS" -eq $Platform) {
         # We're exporting an Xcode project and building that in a separate step.
         ./test/Scripts.Integration.Test/build-project.ps1 -UnityPath "$UnityPath" -UnityVersion $UnityVersion -Platform $Platform
-        & "./scripts/compile-xcode-project.ps1"
+        & (Join-Path $PSScriptRoot "../../scripts/compile-xcode-project.ps1")
     }
     Else {
         ./test/Scripts.Integration.Test/build-project.ps1 -UnityPath "$UnityPath" -UnityVersion $UnityVersion -Platform $Platform