+++ test/Sentry.Unity.Editor.Tests/SentryScriptableObjectTests.cs
@@ -2,45 +2,52 @@ using System;
 using System.IO;
 using NUnit.Framework;
 using UnityEditor;
+using UnityEngine;
 
 namespace Sentry.Unity.Editor.Tests;
 
 public class SentryScriptableObjectTests
 {
-    private string _tempPath = null!; // Assigned in Setup
+    private string _assetPath = null!;
+    private string _filePath = null!;
 
     [SetUp]
-    public void Setup() => _tempPath = Path.Combine("Assets", Guid.NewGuid().ToString(), "TestOptions.asset");
+    public void Setup()
+    {
+        var directoryName = Guid.NewGuid().ToString("N");
+        _assetPath = Path.Combine("Assets", directoryName, "TestOptions.asset");
+        _filePath = Path.Combine(Application.dataPath, directoryName, "TestOptions.asset");
+    }
 
     [TearDown]
-    public void TearDown() => AssetDatabase.DeleteAsset(Path.GetDirectoryName(_tempPath));
+    public void TearDown() => AssetDatabase.DeleteAsset(Path.GetDirectoryName(_assetPath));
 
     [Test]
     public void CreateOrLoad_ScriptableSentryUnityOptionsAssetDoesNotExist_CreatesNewOptionsAsset()
     {
-        Assert.IsFalse(File.Exists(_tempPath)); // Sanity check
+        Assert.IsFalse(File.Exists(_filePath)); // Sanity check
 
-        SentryScriptableObject.CreateOrLoad<ScriptableSentryUnityOptions>(_tempPath);
+        SentryScriptableObject.CreateOrLoad<ScriptableSentryUnityOptions>(_assetPath);
 
-        Assert.IsTrue(File.Exists(_tempPath));
+        Assert.IsTrue(File.Exists(_filePath));
     }
 
     [Test]
     public void CreateOrLoad_SentryCliOptionsAssetDoesNotExist_CreatesNewOptionsAsset()
     {
-        Assert.IsFalse(File.Exists(_tempPath)); // Sanity check
+        Assert.IsFalse(File.Exists(_filePath)); // Sanity check
 
-        SentryScriptableObject.CreateOrLoad<SentryCliOptions>(_tempPath);
+        SentryScriptableObject.CreateOrLoad<SentryCliOptions>(_assetPath);
 
-        Assert.IsTrue(File.Exists(_tempPath));
+        Assert.IsTrue(File.Exists(_filePath));
     }
 
     [Test]
     public void Load_OptionsAssetDoesNotExist_ReturnsNull()
     {
-        Assert.IsFalse(File.Exists(_tempPath)); // Sanity check
+        Assert.IsFalse(File.Exists(_filePath)); // Sanity check
 
-        var options = SentryScriptableObject.Load<ScriptableSentryUnityOptions>(_tempPath);
+        var options = SentryScriptableObject.Load<ScriptableSentryUnityOptions>(_assetPath);
 
         Assert.IsNull(options);
     }
@@ -49,13 +56,13 @@ public class SentryScriptableObjectTests
     public void Load_ScriptableSentryUnityOptionsExist_LoadsSavedOptionsAsset()
     {
         var expectedDsn = "test_dsn";
-        var options = SentryScriptableObject.CreateOrLoad<ScriptableSentryUnityOptions>(_tempPath);
+        var options = SentryScriptableObject.CreateOrLoad<ScriptableSentryUnityOptions>(_assetPath);
         options.Dsn = expectedDsn;
         AssetDatabase.SaveAssets(); // Saving to disk
 
-        Assert.IsTrue(File.Exists(_tempPath)); // Sanity check
+        Assert.IsTrue(File.Exists(_filePath)); // Sanity check
 
-        var actualOptions = SentryScriptableObject.Load<ScriptableSentryUnityOptions>(_tempPath);
+        var actualOptions = SentryScriptableObject.Load<ScriptableSentryUnityOptions>(_assetPath);
 
         Assert.NotNull(actualOptions);
         Assert.AreEqual(expectedDsn, actualOptions!.Dsn);
@@ -65,13 +72,13 @@ public class SentryScriptableObjectTests
     public void Load_SentryCliOptionsExist_LoadsSavedOptionsAsset()
     {
         var expectedAuth = "test_auth";
-        var options = SentryScriptableObject.CreateOrLoad<SentryCliOptions>(_tempPath);
+        var options = SentryScriptableObject.CreateOrLoad<SentryCliOptions>(_assetPath);
         options.Auth = expectedAuth;
         AssetDatabase.SaveAssets(); // Saving to disk
 
-        Assert.IsTrue(File.Exists(_tempPath)); // Sanity check
+        Assert.IsTrue(File.Exists(_filePath)); // Sanity check
 
-        var actualOptions = SentryScriptableObject.Load<SentryCliOptions>(_tempPath);
+        var actualOptions = SentryScriptableObject.Load<SentryCliOptions>(_assetPath);
 
         Assert.NotNull(actualOptions);
         Assert.AreEqual(expectedAuth, actualOptions!.Auth);