+++ test/Sentry.Unity.Tests/SentryUnityOptionsTests.cs
@@ -1,4 +1,6 @@
+using System.Linq;
 using NUnit.Framework;
+using Sentry.Unity.Integrations;
 using Sentry.Unity.Tests.Stubs;
 using UnityEngine;
 
@@ -55,6 +57,10 @@ public sealed class SentryUnityOptionsTests
 
     [Test]
     public void Ctor_IsGlobalModeEnabled_IsTrue() => Assert.IsTrue(_fixture.GetSut().IsGlobalModeEnabled);
+
+    [Test]
+    public void Ctor_CSharpAnrWatchdog_IsNotRegistered() =>
+        Assert.That(_fixture.GetSut().Integrations.OfType<AnrIntegration>(), Is.Empty);
 
     [Test]
     public void Ctor_Release_IsProductNameAtVersion() =>
@@ -158,17 +164,28 @@ public sealed class SentryUnityOptionsTests
     }
 
     [Test]
-    public void Options_Experimental_EnableNativeAppHangTracking_DefaultsToFalse()
+    public void Options_EnableAppHangTracking_DefaultsToFalse()
     {
         var options = new SentryUnityOptions();
-        Assert.IsFalse(options.Experimental.EnableNativeAppHangTracking);
+        Assert.IsFalse(options.EnableAppHangTracking);
     }
 
     [Test]
-    public void Options_Experimental_EnableNativeAppHangTracking_IsSettable()
+    public void Options_EnableAppHangTracking_IsSettable()
     {
         var options = new SentryUnityOptions();
+        options.EnableAppHangTracking = true;
+        Assert.IsTrue(options.EnableAppHangTracking);
+    }
+
+    [Test]
+    public void Options_LegacyNativeAppHangTracking_EnablesNativeAppHangTracking()
+    {
+        var options = new SentryUnityOptions();
+#pragma warning disable CS0618 // Type or member is obsolete
         options.Experimental.EnableNativeAppHangTracking = true;
-        Assert.IsTrue(options.Experimental.EnableNativeAppHangTracking);
+#pragma warning restore CS0618 // Type or member is obsolete
+
+        Assert.IsTrue(options.NativeAppHangTrackingEnabled);
     }
 }